backend/requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Deprecated==1.3.1
et_xmlfile==2.0.0
fastapi==0.128.0
fastapi-socketio==0.0.10
filelock==3.20.3
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
limits==5.8.0
Markdown==3.10
markdown-it-py==4.0.0
markdown2==2.4.10
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.15.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
opt_einsum==3.4.0
optree==0.18.0
ordered-set==4.1.0
orjson==3.11.5
packaging==25.0
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
pscript==0.7.7
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.6
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.0
sentencepiece==0.2.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vbuild==0.8.2
watchfiles==0.24.0
websockets==16.0
Werkzeug==3.1.5
wrapt==2.0.1
wsproto==1.3.2
